Move functional interfaces in IOUtils to top-level interfaces
