+++ lucene/core/src/java/org/apache/lucene/index/BufferedUpdatesStream.java
@@ -29,6 +29,7 @@ import java.util.concurrent.atomic.AtomicLong;
 import org.apache.lucene.store.IOContext;
 import org.apache.lucene.util.Accountable;
 import org.apache.lucene.util.BytesRef;
+import org.apache.lucene.util.IOConsumer;
 import org.apache.lucene.util.IOUtils;
 import org.apache.lucene.util.InfoStream;
 
@@ -280,16 +281,14 @@ final class BufferedUpdatesStream implements Accountable {
     final ReadersAndUpdates rld;
     final SegmentReader reader;
     final int startDelCount;
-    private final IOUtils.IOConsumer<ReadersAndUpdates> onClose;
+    private final IOConsumer<ReadersAndUpdates> onClose;
 
     TermsEnum termsEnum;
     PostingsEnum postingsEnum;
     BytesRef term;
 
     SegmentState(
-        ReadersAndUpdates rld,
-        IOUtils.IOConsumer<ReadersAndUpdates> onClose,
-        SegmentCommitInfo info)
+        ReadersAndUpdates rld, IOConsumer<ReadersAndUpdates> onClose, SegmentCommitInfo info)
         throws IOException {
       this.rld = rld;
       reader = rld.getReader(IOContext.READ);
