+++ lucene/core/src/java/org/apache/lucene/index/IndexWriter.java
@@ -80,6 +80,8 @@ import org.apache.lucene.util.Bits;
 import org.apache.lucene.util.BytesRef;
 import org.apache.lucene.util.Constants;
 import org.apache.lucene.util.Counter;
+import org.apache.lucene.util.IOConsumer;
+import org.apache.lucene.util.IOFunction;
 import org.apache.lucene.util.IOUtils;
 import org.apache.lucene.util.InfoStream;
 import org.apache.lucene.util.StringHelper;
@@ -525,7 +527,7 @@ public class IndexWriter
     final Map<String, SegmentReader> openedReadOnlyClones = new HashMap<>();
     // this function is used to control which SR are opened in order to keep track of them
     // and to reuse them in the case we wait for merges in this getReader call.
-    IOUtils.IOFunction<SegmentCommitInfo, SegmentReader> readerFactory =
+    IOFunction<SegmentCommitInfo, SegmentReader> readerFactory =
         sci -> {
           final ReadersAndUpdates rld = getPooledInstance(sci, true);
           try {
@@ -3569,7 +3571,7 @@ public class IndexWriter
       SegmentInfos mergingSegmentInfos,
       BooleanSupplier stopCollectingMergeResults,
       MergeTrigger trigger,
-      IOUtils.IOConsumer<SegmentCommitInfo> mergeFinished)
+      IOConsumer<SegmentCommitInfo> mergeFinished)
       throws IOException {
     assert Thread.holdsLock(this);
     assert trigger == MergeTrigger.GET_READER || trigger == MergeTrigger.COMMIT
@@ -3664,8 +3666,7 @@ public class IndexWriter
 
                       @Override
                       void initMergeReaders(
-                          IOUtils.IOFunction<SegmentCommitInfo, MergePolicy.MergeReader>
-                              readerFactory)
+                          IOFunction<SegmentCommitInfo, MergePolicy.MergeReader> readerFactory)
                           throws IOException {
                         if (onlyOnce.compareAndSet(false, true)) {
                           // we do this only once below to pull readers as point in time readers
@@ -5545,7 +5546,7 @@ public class IndexWriter
       TrackingDirectoryWrapper directory,
       final SegmentInfo info,
       IOContext context,
-      IOUtils.IOConsumer<Collection<String>> deleteFiles)
+      IOConsumer<Collection<String>> deleteFiles)
       throws IOException {
 
     // maybe this check is not needed, but why take the risk?
