+++ lucene/core/src/test/org/apache/lucene/store/BaseDataOutputTestCase.java
@@ -29,7 +29,7 @@ import java.util.List;
 import java.util.Random;
 import org.apache.lucene.tests.util.LuceneTestCase;
 import org.apache.lucene.util.ArrayUtil;
-import org.apache.lucene.util.IOUtils.IOConsumer;
+import org.apache.lucene.util.IOConsumer;
 import org.junit.Test;
 
 public abstract class BaseDataOutputTestCase<T extends DataOutput> extends LuceneTestCase {
